Deleted collapsible node card component and moved functionalities to detailed node card. Margin/invisible container needs fixing, Resizer seems weird, container size doesn't adapt to node footer size correctly and double click preventing zoom on nodes is broken and needs fixing.
