+++ frontend/app/components/PxComponents/PxNodeCardDetailed.vue
@@ -1,16 +1,20 @@
 <script setup lang="ts">
 import { PxComponentCreationForm, PxNodeFixModal } from '#components'
+import type { ContextMenuItem } from '#ui/components/ContextMenu.vue'
 
 const props = defineProps<{
   node: PxNode
+  isCollapsible: boolean
 }>()
 
 const emit = defineEmits<{
   (e: 'update', updatedNode: PxNode): void
   (e: 'delete', nodeId: string): void
   (e: 'deleteComponent' | 'addComponent', nodeId: string, componentId: string): void
+  (e: 'deleteContainer' | 'switchNode'): void
 }>()
 
+const isCollapsed = ref(true)
 const isBeingEdited = ref(false)
 const toast = useToast()
 const pxNodesLLM = usePxNodesLLM()
@@ -23,13 +27,52 @@ const editForm = ref({
   description: props.node.description,
 })
 
+const menuItems = ref<ContextMenuItem[]>([
+  {
+    label: 'Delete node',
+    onSelect() {
+      handleDeleteContainer()
+    },
+  },
+  {
+    label: 'Switch node',
+    onSelect() {
+      handleSwitchNode()
+    },
+  },
+  {
+    label: 'Edit node',
+    onSelect() {
+      startEdit()
+    },
+  },
+  {
+    label: 'Add component',
+    onSelect() {
+      handleAddComponent()
+    },
+  },
+])
+
 // LLM Feedback state
 const llmFeedback = ref<NodeValidationFeedback | null>(null)
 const isValidating = ref(false)
 
+async function toggleCollapsed() {
+  isCollapsed.value = !isCollapsed.value
+}
+
+function onDbClick(mouseEvent: MouseEvent) {
+  mouseEvent.preventDefault()
+  if (!isBeingEdited.value && !llmFeedback.value) {
+    startEdit()
+  }
+}
+
 function startEdit() {
   editForm.value = { ...props.node }
   isBeingEdited.value = true
+  isCollapsed.value = false
 }
 
 function confirmEdit() {
@@ -58,6 +101,14 @@ async function handleAddComponent() {
   }
 
   emit('addComponent', result.nodeId, result.componentId)
+}
+
+function handleDeleteContainer() {
+  emit('deleteContainer')
+}
+
+async function handleSwitchNode() {
+  emit('switchNode')
 }
 
 // LLM Validation
@@ -117,131 +168,176 @@ async function openFixModal() {
 </script>
 
 <template>
-  <UCard class="hover:shadow-lg transition">
-    <template #header>
-      <h2 v-if="!isBeingEdited" class="font-semibold text-lg">
-        <NuxtLink :to="{ name: 'pxnodes-id', params: { id: props.node.id } }">
-          {{ props.node.name }}
-        </NuxtLink>
-      </h2>
-      <UTextarea v-else v-model="editForm.name" />
-    </template>
-
-    <template #default>
-      <div v-if="!isBeingEdited">
-        <h2 class="font-semibold text-lg mb-2">Description</h2>
-        <p>{{ props.node.description }}</p>
-        <br />
-        <h2 v-if="node.components.length === 0" class="italic">This node has no components.</h2>
-        <h2 v-else class="font-semibold text-lg mb-2">Components</h2>
-        <section class="grid grid-cols-1 gap-6">
-          <div v-for="component in node.components" :key="component.id">
-            <PxComponentCard
-              visualization-style="preview"
-              :component="component"
-              @delete="handleDeleteComponent"
-            />
-          </div>
-        </section>
-        <br />
-        <h2 v-if="node.charts.length === 0" class="italic">
-          This node is not associated to any charts.
+  <UContextMenu :items="menuItems" :disabled="!!(isBeingEdited || llmFeedback || !isCollapsible)">
+    <UCard class="hover:shadow-lg transition" @dblclick="onDbClick($event)">
+      <template #header>
+        <h2 v-if="!isBeingEdited" class="font-semibold text-lg">
+          <NuxtLink :to="{ name: 'pxnodes-id', params: { id: props.node.id } }">
+            {{ props.node.name }}
+          </NuxtLink>
         </h2>
-        <div v-else>
-          <h2 class="font-semibold text-lg mb-2">Associated Charts</h2>
+        <UTextarea v-else v-model="editForm.name" />
+      </template>
+
+      <template #default>
+        <div v-if="!isBeingEdited">
+          <h2 class="font-semibold text-lg mb-2">Description</h2>
+          <p>{{ props.node.description }}</p>
+          <br />
+          <h2 v-if="node.components.length === 0" class="italic">This node has no components.</h2>
+          <h2 v-else class="font-semibold text-lg mb-2">Components</h2>
           <section class="grid grid-cols-1 gap-6">
-            <div v-for="chart in node.charts" :key="chart.id">
-              <PxChartCard :px-chart="chart" :visualization-style="'preview'" />
+            <div v-for="component in node.components" :key="component.id">
+              <PxComponentCard
+                visualization-style="preview"
+                :component="component"
+                @delete="handleDeleteComponent"
+              />
             </div>
           </section>
-        </div>
-
-        <!-- LLM Feedback Section -->
-        <div v-if="llmFeedback" class="mt-6 pt-4 border-t">
-          <!-- Coherence Score -->
-          <div class="mb-3">
-            <div class="flex items-center gap-2 mb-2">
-              <span class="text-sm text-gray-600 dark:text-gray-400">Coherence Score:</span>
-              <UBadge
-                :color="
-                  llmFeedback.overall_coherence_score >= 7
-                    ? 'success'
-                    : llmFeedback.overall_coherence_score >= 4
-                      ? 'warning'
-                      : 'error'
-                "
-                variant="solid"
-              >
-                {{ llmFeedback.overall_coherence_score }}/10
-              </UBadge>
-            </div>
-            <p class="text-sm text-gray-600 dark:text-gray-300">{{ llmFeedback.summary }}</p>
+          <br />
+          <h2 v-if="node.charts.length === 0" class="italic">
+            This node is not associated to any charts.
+          </h2>
+          <div v-else>
+            <h2 class="font-semibold text-lg mb-2">Associated Charts</h2>
+            <section class="grid grid-cols-1 gap-6">
+              <div v-for="chart in node.charts" :key="chart.id">
+                <PxChartCard :px-chart="chart" :visualization-style="'preview'" />
+              </div>
+            </section>
           </div>
 
-          <!-- Show issues list -->
-          <div v-for="(issue, index) in llmFeedback?.issues" :key="index">
-            <UAlert
-              class="mb-2"
-              variant="subtle"
-              :color="issue.severity >= 3 ? 'error' : 'warning'"
-              :title="issue.title"
-              :description="`${issue.description} (Severity: ${issue.severity})`"
-              :actions="[
-                {
-                  label: 'Dismiss',
-                  color: 'warning',
-                  variant: 'subtle',
-                  class: 'ml-auto',
-                  onClick: () => dismissIssue(index),
-                },
-              ]"
-            />
-            <!-- Related components -->
-            <div v-if="issue.related_components.length > 0" class="ml-4 mb-2 flex flex-wrap gap-1">
-              <span class="text-xs text-gray-500">Related components:</span>
-              <UBadge
-                v-for="comp in issue.related_components"
-                :key="comp"
-                color="neutral"
+          <!-- LLM Feedback Section -->
+          <div v-if="llmFeedback" class="mt-6 pt-4 border-t">
+            <!-- Coherence Score -->
+            <div class="mb-3">
+              <div class="flex items-center gap-2 mb-2">
+                <span class="text-sm text-gray-600 dark:text-gray-400">Coherence Score:</span>
+                <UBadge
+                  :color="
+                    llmFeedback.overall_coherence_score >= 7
+                      ? 'success'
+                      : llmFeedback.overall_coherence_score >= 4
+                        ? 'warning'
+                        : 'error'
+                  "
+                  variant="solid"
+                >
+                  {{ llmFeedback.overall_coherence_score }}/10
+                </UBadge>
+              </div>
+              <p class="text-sm text-gray-600 dark:text-gray-300">{{ llmFeedback.summary }}</p>
+            </div>
+
+            <!-- Show issues list -->
+            <div v-for="(issue, index) in llmFeedback?.issues" :key="index">
+              <UAlert
+                class="mb-2"
                 variant="subtle"
-                size="xs"
+                :color="issue.severity >= 3 ? 'error' : 'warning'"
+                :title="issue.title"
+                :description="`${issue.description} (Severity: ${issue.severity})`"
+                :actions="[
+                  {
+                    label: 'Dismiss',
+                    color: 'warning',
+                    variant: 'subtle',
+                    class: 'ml-auto',
+                    onClick: () => dismissIssue(index),
+                  },
+                ]"
+              />
+              <!-- Related components -->
+              <div
+                v-if="issue.related_components.length > 0"
+                class="ml-4 mb-2 flex flex-wrap gap-1"
               >
-                {{ comp }}
-              </UBadge>
+                <span class="text-xs text-gray-500">Related components:</span>
+                <UBadge
+                  v-for="comp in issue.related_components"
+                  :key="comp"
+                  color="neutral"
+                  variant="subtle"
+                  size="xs"
+                >
+                  {{ comp }}
+                </UBadge>
+              </div>
+            </div>
+
+            <!-- Single "Fix All Issues" button (only show if there are issues) -->
+            <UButton
+              v-if="(llmFeedback?.issues?.length ?? 0) > 0"
+              class="mt-3 w-full"
+              color="primary"
+              icon="i-heroicons-sparkles"
+              label="Fix All Issues with AI"
+              @click="openFixModal"
+            />
+          </div>
+        </div>
+        <UTextarea v-else v-model="editForm.description" />
+      </template>
+
+      <template #footer>
+        <div v-if="!isBeingEdited">
+          <!--collapsed-->
+          <div
+            v-if="isCollapsed && isCollapsible"
+            class="flex justify-center"
+            style="padding: -5px; margin: -5px; cursor: default"
+            @click="toggleCollapsed()"
+          >
+            <Icon name="heroicons:chevron-down-20-solid" class="size-5" />
+          </div>
+          <!-- not collapsed-->
+          <div v-else class="flex-col justify-between">
+            <div class="flex justify-between gap-2">
+              <UButton
+                color="warning"
+                variant="soft"
+                :loading="isValidating"
+                @click="handleValidation"
+              >
+                Check
+              </UButton>
+              <div class="flex gap-2">
+                <UButton color="primary" variant="soft" @click="handleAddComponent">
+                  Add Component
+                </UButton>
+                <UButton color="secondary" variant="soft" @click="startEdit">Edit</UButton>
+                <UButton
+                  v-if="isCollapsible"
+                  color="secondary"
+                  variant="soft"
+                  @click="handleSwitchNode()"
+                  >Switch Node</UButton
+                >
+                <UButton v-if="!isCollapsible" color="error" variant="soft" @click="emitDelete"
+                  >Delete</UButton
+                >
+                <UButton v-else color="error" variant="soft" @click="handleDeleteContainer()"
+                  >Delete</UButton
+                >
+              </div>
+            </div>
+            <div
+              v-if="isCollapsible"
+              class="flex justify-center"
+              style="padding-top: 15px; cursor: default"
+              @click="toggleCollapsed()"
+            >
+              <Icon name="heroicons:chevron-up-20-solid" class="size-5" />
             </div>
           </div>
-
-          <!-- Single "Fix All Issues" button (only show if there are issues) -->
-          <UButton
-            v-if="(llmFeedback?.issues?.length ?? 0) > 0"
-            class="mt-3 w-full"
-            color="primary"
-            icon="i-heroicons-sparkles"
-            label="Fix All Issues with AI"
-            @click="openFixModal"
-          />
         </div>
-      </div>
-      <UTextarea v-else v-model="editForm.description" />
-    </template>
-
-    <template #footer>
-      <div v-if="!isBeingEdited" class="flex justify-between">
-        <UButton color="warning" variant="soft" :loading="isValidating" @click="handleValidation">
-          Check
-        </UButton>
-        <div class="flex gap-2">
-          <UButton color="primary" variant="soft" @click="handleAddComponent">
-            Add Component
-          </UButton>
-          <UButton color="secondary" variant="soft" @click="startEdit">Edit</UButton>
-          <UButton color="error" variant="soft" @click="emitDelete">Delete</UButton>
+        <!-- during edit-->
+        <div v-else class="flex gap-2">
+          <UButton color="error" variant="soft" @click="cancelEdit">Cancel</UButton>
+          <UButton color="secondary" variant="soft" @click="confirmEdit">Confirm</UButton>
         </div>
-      </div>
-      <div v-else class="flex gap-2">
-        <UButton color="error" variant="soft" @click="cancelEdit">Cancel</UButton>
-        <UButton color="secondary" variant="soft" @click="confirmEdit">Confirm</UButton>
-      </div>
-    </template>
-  </UCard>
+      </template>
+    </UCard>
+  </UContextMenu>
 </template>
